### 2026-01-30 #9919 @danielskatz
👋 @svretina - Thanks for your submission. Please add the other required sections to your paper.

Please see:

* Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
* https://joss.readthedocs.io/en/latest/submitting.html#submitting-a-paper-to-joss
* An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html
