### 2026-01-30 7676c1a
add JOSS badge to README
