+++ README.md
@@ -8,6 +8,8 @@
 [![Coverage](https://codecov.io/gh/svretina/FastTanhSinhQuadrature.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/svretina/FastTanhSinhQuadrature.jl)
 [![Aqua QA](https://raw.githubusercontent.com/JuliaTesting/Aqua.jl/master/badge.svg)](https://github.com/JuliaTesting/Aqua.jl)
 [![](https://img.shields.io/badge/%F0%9F%9B%A9%EF%B8%8F_tested_with-JET.jl-233f9a)](https://github.com/aviatesk/JET.jl)
+[![status](https://joss.theoj.org/papers/84c42780ad01b47c0c3b5f1ab5c26260/status.svg)](https://joss.theoj.org/papers/84c42780ad01b47c0c3b5f1ab5c26260)
+
 
 Fast and high-precision numerical integration using **Tanh-Sinh (Double Exponential) quadrature** in Julia.
 
