### 2026-01-30 74ecb82
adds missing sections to JOSS paper
