+++ JOSS_paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-Numerical integration is a cornerstone of scientific computing, essential for evaluating integrals that cannot be solved analytically. The Tanh-Sinh (or Double Exponential) quadrature, originally proposed by Takahasi and Mori (@Takahasi1973), is a powerful technique known for its high accuracy and efficiency, particularly for integrands with endpoint singularities. `FastTanhSinhQuadrature.jl` provides a high-performance, arbitrary-precision implementation of this method in Julia. It leverages modern compiler technologies to achieve significant speedups over traditional implementations while maintaining rigorous mathematical precision.
+Numerical integration is a cornerstone of scientific computing, essential for evaluating integrals that cannot be solved analytically. The Tanh-Sinh (or Double Exponential) quadrature, originally proposed by @Takahasi1973, is a powerful technique known for its high accuracy and efficiency, particularly for integrands with endpoint singularities. `FastTanhSinhQuadrature.jl` provides a high-performance, arbitrary-precision implementation of this method in Julia. It leverages modern compiler technologies to achieve significant speedups over traditional implementations while maintaining rigorous mathematical precision.
 
 # Statement of Need
 
@@ -31,6 +31,93 @@ In many fields of physics and engineering, researchers encounter integrals with 
 3.  **Flexibility**: Supporting arbitrary precision types (e.g., `BigFloat`) and multidimensional integration.
 
 `FastTanhSinhQuadrature.jl` addresses these needs by implementing a rigorous Tanh-Sinh scheme with an optimized "window selection" strategy enabling SIMD-accelerated execution paths. This makes it an ideal tool for large-scale simulations where both speed and precision are critical.
+
+# State of the field
+
+Numerical integration is a widely solved problem, with mature implementations available in libraries such as **Boost** (C++), **SciPy** (Python), and **mpmath** (Python). Within the Julia ecosystem, packages like `QuadGK.jl` and `HCubature.jl` provide efficient adaptive Gauss-Kronrod and h-adaptive cubature methods, respectively. However, implementations specifically of the **Tanh-Sinh quadrature** are less common and often face a fundamental performance bottleneck.
+
+Most existing Tanh-Sinh implementations, both in Julia and other languages, rely on **dynamic condition checks** within the summation loop. These algorithms typically iterate until the current term drops below a certain tolerance to handle the infinite domain truncation and detect underflow. While mathematically sound, this approach introduces conditional branching inside the innermost loop, which prevents modern compilers from applying **Single Instruction, Multiple Data (SIMD)** vectorization. Consequently, these solvers are restricted to scalar execution speeds.
+
+`FastTanhSinhQuadrature.jl` was developed to bridge this gap. Rather than modifying existing scalar implementations, we built a solver from the ground up centered on the **"window selection"** strategy described by @Vanherck2020. This method analytically pre-calculates the optimal step size $h$ and the truncation index $n$ based strictly on the floating-point precision (e.g., `Float64`, `BigFloat`). By determining the bounds *a priori*, we eliminate the need for runtime convergence checks. This creates a branch-free inner loop that allows the compiler (specifically via `LoopVectorization.jl`) to generate highly optimized SIMD instructions.
+
[…]
