### 2026-01-30 #9919 @svretina
Possible reviewers:
@ranocha (https://reviewers.joss.theoj.org/reviewers/1437)
Tom Kimpson (https://reviewers.joss.theoj.org/reviewers/3425)
Emerson Martins de Andrade (https://reviewers.joss.theoj.org/reviewers/3241)
Xuanzhao Gao (https://reviewers.joss.theoj.org/reviewers/25168)
