### 2026-02-03 1339ff5
improve adaptive integration: remove weight/node calculation inside of loop
