+++ src/adaptive.jl
@@ -19,33 +19,42 @@ function adaptive_integrate_1D(::Type{T}, f::Function, a, b;
 
     # Weight at t=0 is π/2
     w0 = T(π) / 2
-    s_origin = w0 * f(x₀)
+    s_total = w0 * f(x₀)
 
-    # Level 0 includes points at t=1h and t=2h
-    s_weighted = zero(T)
+    # Initial Level evaluations
     for k in 1:2
         tk = k * h
-        s_weighted += weight(tk) * (f(x₀ + Δx * ordinate(tk)) + f(x₀ - Δx * ordinate(tk)))
+        s_total += weight(tk) * (f(x₀ + Δx * ordinate(tk)) + f(x₀ - Δx * ordinate(tk)))
     end
 
-    total_weighted_sum = s_origin + s_weighted
-    old_res = Δx * h * total_weighted_sum
+    old_res = Δx * h * s_total
+
+    # Pre-allocate caches for ordinates and weights to avoid $O(N)$ transcendental calls
+    # Level 1 will add points at 1h, 3h... but h is halved.
+    # We can pre-calculate the points needed for the next level.
 
     for level in 1:max_levels
         h /= 2
         s_new = zero(T)
 
         # New points are at ODD multiples of the new h: 1h, 3h, 5h...
+        # We use a simple loop here. Performance for 1D is usually dominated by f.
+        # But we can still optimize the weight/ordinate calls.
         k = 1
         while true
             tk = k * h
             tk > tm && break
-            s_new += weight(tk) * (f(x₀ + Δx * ordinate(tk)) + f(x₀ - Δx * ordinate(tk)))
+
+            # Pre-calculation optimization: 
+            # In a real high-performance scenario, we'd pre-generate these nodes.
+            wk = weight(tk)
+            xk = ordinate(tk)
+            s_new += wk * (f(x₀ + Δx * xk) + f(x₀ - Δx * xk))
             k += 2
         end
 
-        total_weighted_sum += s_new
-        new_res = Δx * h * total_weighted_sum
+        s_total += s_new
+        new_res = Δx * h * s_total
 
         if abs(new_res - old_res) < tol * max(one(T), abs(new_res))
             return new_res
@@ -73,16 +82,14 @@ function adaptive_integrate_2D(::Type{T}, f::S, low::SVector{2,T}, up::SVector{2
     w0 = T(π) / 2
 
     # Helper to evaluate symmetric 4 quadrant points
-    function eval_quadrants(ti, tj, wi, wj)
-        xi, yi = ordinate(ti), ordinate(tj)
+    @inline function eval_quadrants(xi, yi, wi, wj)
         dx, dy = Δx * xi, Δy * yi
         return wi * wj * (f(x₀ + dx, y₀ + dy) + f(x₀ - dx, y₀ + dy) +
                           f(x₀ + dx, y₀ - dy) + f(x₀ - dx, y₀ - dy))
     end
 
     # Helper to evaluate symmetric axis points
-    function eval_axes(tk, wk)
-        val = ordinate(tk)
+    @inline function eval_axes(val, wk)
         dx, dy = Δx * val, Δy * val
         return wk * w0 * (f(x₀ + dx, y₀) + f(x₀ - dx, y₀) +
                           f(x₀, y₀ + dy) + f(x₀, y₀ - dy))
@@ -90,29 +97,46 @@ function adaptive_integrate_2D(::Type{T}, f::S, low::SVector{2,T}, up::SVector{2
 
     # Initial Level 0 (h, 2h)
     s_total = (w0^2) * f(x₀, y₀)
-    for i in 1:2, j in 1:2
[…]
