+++ Project.toml
@@ -9,7 +9,16 @@ LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
+DoubleFloats = "1.1, 1.2"
 LambertW = "1"
 LoopVectorization = "0.12"
 StaticArrays = "1"
+Test = "1"
 julia = "1.10, 1.11"
+
+[extras]
+DoubleFloats = "497a8b3b-efae-58df-a0af-a86822472b78"
+Test = "8dfed614-2900-4131-9a72-6804100913c7"
+
+[targets]
+test = ["DoubleFloats", "Test"]

[…]
