### 2026-02-04 #9919 @danielskatz
@svretina - Can you reduce the size of your paper? JOSS now suggests papers be between 750 and 1750 words, with the new sections.  Yours is about 600 words over this max.

Perhaps either the Performance benchmarks or Math could be summarized, and could point to something in the repo or docs for more of the details?

You might also add a blank link in your .md file before the list in the Statement of Need section to get it to appear as a list.

Finally, you might add references to the packages you mention in the State of the field section.
