### 2026-02-05 d25ebe2
docs(JOSS): revise paper for word count and add missing citations
