+++ JOSS_paper/paper.bib
@@ -18,3 +18,66 @@
   url     = {https://arxiv.org/abs/2007.15057},
   doi     = {10.48550/arXiv.2007.15057}
 }
+
+@article{Virtanen2020,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and {van der Walt}, St{\'e}fan J. and Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C J and Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and Harris, Charles R. and Archibald, Anne M. and Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
+  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in Python}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  adsurl  = {https://rdcu.be/b08Wh},
+  doi     = {10.1038/s41592-019-0686-2}
+}
+
+@misc{Boost,
+  author = {{Boost C++ Libraries}},
+  title  = {Boost C++ Libraries},
+  url    = {https://www.boost.org/},
+  year   = {2024}
+}
+
+@misc{Johansson_mpmath,
+  author = {Johansson, Fredrik and others},
+  title  = {mpmath: a {Python} library for arbitrary-precision floating-point arithmetic},
+  url    = {http://mpmath.org/},
+  year   = {2023}
+}
+
+@misc{QuadGK,
+  author = {Johnson, Steven G.},
+  title  = {QuadGK.jl: Gauss-Kronrod integration in Julia},
+  url    = {https://github.com/JuliaMath/QuadGK.jl},
+  year   = {2013}
+}
+
+@misc{HCubature,
+  author = {Johnson, Steven G.},
+  title  = {HCubature.jl: H-adaptive multidimensional integration in Julia},
+  url    = {https://github.com/JuliaMath/HCubature.jl},
+  year   = {2017}
+}
+
+@article{LoopVectorization,
+  author    = {Chris Elrod},
+  title     = {LoopVectorization.jl},
+  journal   = {Zenodo},
+  year      = {2022},
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.7303032},
+  url       = {https://doi.org/10.5281/zenodo.7303032}
+}
+
+@misc{FastGaussQuadrature,
+  author = {Townsend, Alex and Hale, Nicholas and Olver, Sheehan},
+  title  = {FastGaussQuadrature.jl},
+  url    = {https://github.com/JuliaApproximation/FastGaussQuadrature.jl},
+  year   = {2013}
+}
+
+@misc{Integrals,
+  author = {Widmann, David and Rackauckas, Christopher},
+  title  = {Integrals.jl: A common interface for numerical integration in Julia},
+  url    = {https://github.com/SciML/Integrals.jl},
+  year   = {2020}
+}
