+++ JOSS_paper/paper.md
@@ -25,144 +25,53 @@ Numerical integration is a cornerstone of scientific computing, essential for ev
 
 # Statement of Need
 
-In many fields of physics and engineering, researchers encounter integrals with singularities at the boundaries. Standard Gaussian quadrature rules often fail or require an excessive number of points to converge in these cases. While other integration libraries exist, they don't offer the combination of:
-1.  **Robustness**: Handling singularities automatically without manual coordinate transformations.
+In many fields of physics and engineering, researchers encounter integrals with singularities at the boundaries. Standard Gaussian quadrature rules often fail or require an excessive number of points to converge in these cases. While other integration libraries exist, they typically lack a combination of robustness, performance, and flexibility. `FastTanhSinhQuadrature.jl` addresses these needs by:
+
+1.  **Robustness**: Automatically handling endpoint singularities without manual coordinate transformations.
 2.  **Performance**: Utilizing SIMD (Single Instruction, Multiple Data) instructions for rapid evaluation.
 3.  **Flexibility**: Supporting arbitrary precision types (e.g., `BigFloat`) and multidimensional integration.
 
-`FastTanhSinhQuadrature.jl` addresses these needs by implementing a rigorous Tanh-Sinh scheme with an optimized "window selection" strategy enabling SIMD-accelerated execution paths. This makes it an ideal tool for large-scale simulations where both speed and precision are critical.
+This package implements a rigorous Tanh-Sinh scheme with an optimized "window selection" strategy enabling SIMD-accelerated execution paths, making it ideal for large-scale simulations where both speed and precision are critical.
 
 # State of the field
 
-Numerical integration is a widely solved problem, with mature implementations available in libraries such as **Boost** (C++), **SciPy** (Python), and **mpmath** (Python). Within the Julia ecosystem, packages like `QuadGK.jl` and `HCubature.jl` provide efficient adaptive Gauss-Kronrod and h-adaptive cubature methods, respectively. However, implementations specifically of the **Tanh-Sinh quadrature** are less common and often face a fundamental performance bottleneck.
+Numerical integration is a well-established field, with mature implementations in libraries such as **Boost** [@Boost] (C++), **SciPy** [@Virtanen2020] (Python), and **mpmath** [@Johansson_mpmath] (Python). Within the Julia ecosystem, packages like `QuadGK.jl` [@QuadGK] and `HCubature.jl` [@HCubature] provide detailed adaptive Gauss-Kronrod and h-adaptive cubature methods. However, high-performance implementations specifically of the **Tanh-Sinh quadrature** are less common.
 
-Most existing Tanh-Sinh implementations, both in Julia and other languages, rely on **dynamic condition checks** within the summation loop. These algorithms typically iterate until the current term drops below a certain tolerance to handle the infinite domain truncation and detect underflow. While mathematically sound, this approach introduces conditional branching inside the innermost loop, which prevents modern compilers from applying **Single Instruction, Multiple Data (SIMD)** vectorization. Consequently, these solvers are restricted to scalar execution speeds.
-
[…]
