### 2026-02-05 e6a03d3
fix loopvectorization ref
