+++ JOSS_paper/paper.bib
@@ -59,13 +59,9 @@
 }
 
 @article{LoopVectorization,
-  author    = {Chris Elrod},
-  title     = {LoopVectorization.jl},
-  journal   = {Zenodo},
-  year      = {2022},
-  publisher = {Zenodo},
-  doi       = {10.5281/zenodo.7303032},
-  url       = {https://doi.org/10.5281/zenodo.7303032}
+  author = {Chris Elrod},
+  title  = {LoopVectorization.jl},
+  url    = {https://github.com/JuliaSIMD/LoopVectorization.jl}
 }
 
 @misc{FastGaussQuadrature,
