### 2026-02-09 #9919 @danielskatz
@svretina - note that your current LICENSE file is actually the instructions on how to apply the license. Please see the part start on line 623 for the specific instructions. Most of what you have in the LICENSE file should be removed and replaced with the text below that, with some things filled in.
