### 2026-02-09 0d9a5aa
apply GNU GPL3 LICENSE
