+++ benchmark/Project.toml
@@ -2,3 +2,5 @@
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 FastGaussQuadrature = "442a2c76-b920-505d-bb47-c5924d526838"
 FastTanhSinhQuadrature = "b650e0df-f744-4436-b963-b44034668c57"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+StrideArrays = "d1fa6d79-ef01-42a6-86c9-f7c551f8593b"

[…]
