### 2026-02-10 #9919 @svretina
> [@svretina](https://github.com/svretina) - note that your current LICENSE file is actually the instructions on how to apply the license. Please see the part start on line 623 for the specific instructions. Most of what you have in the LICENSE file should be removed and replaced with the text below that, with some things filled in.

@danielskatz i implemented the License. package and paper are ready for review.
