### 2026-02-19 #9919 @vissarion
Thanks @ranocha, it is ok if the review is not done immediately. The deadline is typically six weeks. Is this OK with you?
