### 2026-02-19 #10076 @editorialbot
[REVIEW]: FastTanhSinhQuadrature.jl: High-performance Tanh-Sinh numerical integration in Julia
**Submitting author:** @svretina (<a href="http://orcid.org/0000-0001-7575-813X">Stamatis Vretinaris</a>)
**Repository:** https://github.com/svretina/FastTanhSinhQuadrature.jl
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.0
