### 2026-02-25 65a86d3
Fix quad_cmpl complementary coordinate handling
