### 2026-02-25 f2eccbf
Accept Real bounds in precomputed 1D integration
