### 2026-02-25 efe7814
Generalize integration bounds to mixed Real types
