+++ src/integrate2D.jl
@@ -80,6 +80,39 @@ function integrate2D(f::S, low::SVector{2,T}, up::SVector{2,T},
 end
 
 """
+    integrate2D(f, low::SVector{2,<:Real}, up::SVector{2,<:Real}, x, w, h)
+
+Calculate the 2D integral of `f` over `[low, up]` using pre-computed nodes/weights.
+Bounds are converted to the node type `T` to support mixed real types such as `π`.
+"""
+function integrate2D(f::S, low::SVector{2,L}, up::SVector{2,U},
+    x::X, w::W, h::T) where {T<:Real,L<:Real,U<:Real,S,X<:AbstractVector{T},W<:AbstractVector{T}}
+    return integrate2D(
+        f,
+        SVector{2,T}(T(low[1]), T(low[2])),
+        SVector{2,T}(T(up[1]), T(up[2])),
+        x, w, h
+    )
+end
+
+"""
+    integrate2D(f, low::AbstractVector{<:Real}, up::AbstractVector{<:Real}, x, w, h)
+
+Convenience overload for bounds given as vectors of length 2.
+"""
+function integrate2D(f::S, low::AbstractVector{<:Real}, up::AbstractVector{<:Real},
+    x::X, w::W, h::T) where {T<:Real,S,X<:AbstractVector{T},W<:AbstractVector{T}}
+    length(low) == 2 || throw(DimensionMismatch("low must have length 2"))
+    length(up) == 2 || throw(DimensionMismatch("up must have length 2"))
+    return integrate2D(
+        f,
+        SVector{2,T}(T(low[1]), T(low[2])),
+        SVector{2,T}(T(up[1]), T(up[2])),
+        x, w, h
+    )
+end
+
+"""
     integrate2D_avx(f, low, up, x, w, h)
 
 SIMD-accelerated 2D integration over `[low, up]` Using `LoopVectorization`.
@@ -117,4 +150,36 @@ function integrate2D_avx(f::S, low::SVector{2,T}, up::SVector{2,T},
         s += wi * inner_s
     end
     return Δx * Δy * h^2 * s
+end
+
+"""
+    integrate2D_avx(f, low::SVector{2,<:Real}, up::SVector{2,<:Real}, x, w, h)
+
+SIMD-accelerated 2D integration where bounds are converted to node type `T`.
+"""
+function integrate2D_avx(f::S, low::SVector{2,L}, up::SVector{2,U},
+    x::X, w::W, h::T) where {T<:Real,L<:Real,U<:Real,S,X<:AbstractVector{T},W<:AbstractVector{T}}
+    return integrate2D_avx(
+        f,
+        SVector{2,T}(T(low[1]), T(low[2])),
+        SVector{2,T}(T(up[1]), T(up[2])),
+        x, w, h
+    )
+end
+
+"""
+    integrate2D_avx(f, low::AbstractVector{<:Real}, up::AbstractVector{<:Real}, x, w, h)
+
+Convenience SIMD overload for bounds given as vectors of length 2.
+"""
+function integrate2D_avx(f::S, low::AbstractVector{<:Real}, up::AbstractVector{<:Real},
+    x::X, w::W, h::T) where {T<:Real,S,X<:AbstractVector{T},W<:AbstractVector{T}}
+    length(low) == 2 || throw(DimensionMismatch("low must have length 2"))
+    length(up) == 2 || throw(DimensionMismatch("up must have length 2"))
+    return integrate2D_avx(
+        f,
+        SVector{2,T}(T(low[1]), T(low[2])),
+        SVector{2,T}(T(up[1]), T(up[2])),
+        x, w, h
+    )
 end
