+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"convergence/#Convergence","page":"Convergence","title":"Convergence","text":"This page demonstrates the convergence properties of Tanh-Sinh quadrature for challenging test functions from the numerical integration literature.","category":"section"},{"location":"convergence/#Reference","page":"Convergence","title":"Reference","text":"The Tanh-Sinh (Double Exponential) quadrature method was introduced by:\n\nDouble Exponential Formulas for Numerical Integration   Hidetosi Takahasi, Masatake Mori   Publ. Res. Inst. Math. Sci. 9 (1973), no. 3, pp. 721–741   DOI: 10.2977/PRIMS/1195192451\n\nThe last two test functions below are taken directly from this foundational paper.","category":"section"},{"location":"convergence/#Convergence-Plot","page":"Convergence","title":"Convergence Plot","text":"The following plot shows how the absolute integration error decreases as the number of quadrature points increases. The test functions include:\n\nlog(1+x) — Logarithmic function with mild endpoint behavior\n1(1+25x^2) — The Runge function, known to be challenging for polynomial-based methods\nfrac1(x-2)(1-x)^14(1+x)^34 — A challenging singular integrand (from Takahasi-Mori)\nfraccos(pi x)sqrt1-x — Oscillatory function with endpoint singularity (from Takahasi-Mori)\n\n(Image: Convergence of Tanh-Sinh Quadrature)","category":"section"},{"location":"convergence/#Key-Observations","page":"Convergence","title":"Key Observations","text":"Double Exponential Convergence: For all test functions, the error decreases exponentially (linear on the log-scale plot) as the number of points increases.\nSingularity Handling: Even for functions with algebraic singularities at endpoints (like (1-x)^14 or sqrt1-x), Tanh-Sinh quadrature maintains excellent convergence.\nHigh Precision: With 256-bit BigFloat precision, the quadrature achieves errors below 10^-60 with only a few hundred points.","category":"section"},{"location":"convergence/#Test-Functions","page":"Convergence","title":"Test Functions","text":"The exact values were computed using arbitrary-precision arithmetic. Here are the test integrals:\n\nFunction Domain Exact Value Source\nlog(1+x) -1 1 2log(2) - 2 approx -06137 Standard\n1(1+25x^2) -1 1 (25)arctan(5) approx 05493 Runge\nfrac1(x-2)(1-x)^14(1+x)^34 -1 1 approx -19491 Takahasi-Mori (1973)\nfraccos(pi x)sqrt1-x -1 1 approx -06905 Takahasi-Mori (1973)","category":"section"},{"location":"convergence/#Reproducing-the-Plot","page":"Convergence","title":"Reproducing the Plot","text":"The convergence plot can be regenerated using the script docs/convergence_plots.jl:\n\ninclude(\"docs/convergence_plots.jl\")\n\nThis requires the packages: FastTanhSinhQuadrature, DoubleFloats, GaussQuadrature, CairoMakie, LaTeXStrings.","category":"section"},{"location":"benchmarks/#Benchmarks","page":"Benchmarks","title":"Benchmarks","text":"Comparison of FastTanh
[…]
