### 2026-02-25 5a202f1
Expand CI matrix through Julia 1.13 and gate AVX failures
