+++ Project.toml
@@ -14,7 +14,7 @@ LambertW = "1"
 LoopVectorization = "0.12"
 StaticArrays = "1"
 Test = "1"
-julia = "1.9, 1.10, 1.11"
+julia = "1.9, 1.10, 1.11, 1.12"
 
 [extras]
 DoubleFloats = "497a8b3b-efae-58df-a0af-a86822472b78"
