### 2026-02-25 78a0363
Split tests into focused files and keep 3D bounded regression fix
