### 2026-02-25 3e9456f
Refresh README and docs for API usage and bound-type support
