+++ README.md
@@ -21,6 +21,7 @@ Fast and high-precision numerical integration using **Tanh-Sinh (Double Exponent
 - **Arbitrary Precision Support**: Seamlessly works with `Float32`, `Float64`, `BigFloat`, and extended precision types like `Double64` (from `DoubleFloats.jl`).
 - **High Performance**: Specialized `integrate1D_avx`, `integrate2D_avx`, `integrate3D_avx` routines utilize `LoopVectorization.jl` for maximum speed.
 - **Multidimensional Support**: Built-in support for **1D**, **2D**, and **3D** integration domains.
+- **Flexible Bound Types**: Pre-computed and high-level interfaces accept mixed `Real` bounds (for example `Int`, `Float64`, and `π`) and convert them safely to the node type.
 - **Memory Efficiency**: Pre-compute quadrature nodes and weights once and reuse them for multiple integrations.
 - **Singularity Handling**: Robust handling of functions with singularities at integration boundaries via `quad_split` and the new **Complement Interface** `quad_cmpl`.
 - **Double Exponential Convergence**: Achieve machine precision with few points even for singular integrands.
@@ -46,6 +47,13 @@ Pkg.add("FastTanhSinhQuadrature")
 Supported Julia versions: `1.9` - `1.12`.
 
 ## Quick Start
+
+### Choosing an Interface
+
+- Use `quad` for one-off integrations and automatic refinement.
+- Use `integrate1D`/`integrate2D`/`integrate3D` with pre-computed `(x, w, h)` when evaluating many integrals on the same geometry.
+- Use `_avx` variants for `Float32`/`Float64` when your integrand is compatible with `LoopVectorization`.
+- Use `quad_split` for interior singularities and `quad_cmpl` for endpoint-sensitive formulas involving `1-x` / `1+x`.
 
 ### High-Level API: `quad`
 
@@ -98,6 +106,7 @@ x, w, h = tanhsinh(Float64, 80)
 f1(x) = sin(x)^2
 f2(x) = cos(x)^2
 
+# Bounds can be any Real type (Float64, Int, Irrational such as π).
 res1 = integrate1D(f1, 0.0, π, x, w, h)
 res2 = integrate1D(f2, 0.0, π, x, w, h)
 println("Integrals: $res1, $res2")  # Both ≈ π/2
@@ -164,9 +173,9 @@ println(val)  # ≈ 1.0
 | Function | Description |
 |----------|-------------|
 | `quad(f; tol, max_levels)` | Adaptive 1D integration over `[-1, 1]` |
-| `quad(f, low, up; tol, max_levels)` | Adaptive 1D integration over `[low, up]` |
+| `quad(f, low, up; tol, max_levels)` | Adaptive 1D integration over `[low, up]` for `low, up <: Real` |
 | `quad_cmpl(f, low, up; ...)` | High-accuracy 1D integration for `f(x, b-x, x-a)` |
-| `quad(f, low, up; ...)` | Adaptive 2D/3D integration (accepts `SVector` bounds) |
+| `quad(f, low, up; ...)` | Adaptive 2D/3D integration (accepts `SVector` or vectors of reals) |
 | `quad_split(f, c; ...)` | Split domain `[-1, 1]` at singularity `c` and integrate |
 | `quad_split(f, c, low, up; ...)` | Split domain `[low, up]` at singularity `c` and integrate |
 
@@ -179,11 +188,11 @@ println(val)  # ≈ 1.0
 | `integrate1D(f, N)` | Integrate `f` over `[-1, 1]` using `N` points |
[…]
