### 2026-02-26 cd93fba
Clarify quad_cmpl docs and add complement-family tests
