+++ src/quad.jl
@@ -268,9 +268,13 @@ end
 """
     quad_cmpl(f, [low, up]; tol=1e-12, max_levels=10)
 
-High-level interface for Tanh-Sinh quadrature with endpoint sensitivity.
-`f` should accept three arguments: `f(x, b-x, x-a)` for interval `[a, b]`.
-For `[-1, 1]`, this is `f(x, 1-x, 1+x)`.
+High-level 1D interface for endpoint-distance-aware integrands.
+Call as `quad_cmpl(f, a, b)`. At each quadrature node `x` in `[a, b]`,
+the callback is evaluated as `f(x, b_minus_x, x_minus_a)`, where
+`b_minus_x = b - x` and `x_minus_a = x - a` (for `[-1, 1]`: `f(x, 1-x, 1+x)`).
+
+This is useful when expressions like `log(b_minus_x)` or
+`1/sqrt(b_minus_x*x_minus_a)` are sensitive to cancellation near endpoints.
 """
 function quad_cmpl(f::Function, low::T, up::T; tol::Real=1e-12, max_levels::Int=10) where {T<:Real}
     if !(T <: AbstractFloat)
