### 2026-02-26 5dc3674
Add multi-library benchmarks and publish timing tables
