+++ benchmark/benchmarks.jl
@@ -1,75 +1,440 @@
-# FastTanhSinhQuadrature.jl: Fast and high-precision numerical integration using Tanh-Sinh (Double Exponential) quadrature.
-# Copyright (C) 2024-2026 Stamatis Vretinaris
-# 
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-# 
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-using FastTanhSinhQuadrature
-using FastGaussQuadrature
 using BenchmarkTools
+using Cuba
+using Cubature
+using FastGaussQuadrature
+using FastTanhSinhQuadrature
+using HCubature
 using Printf
-using LinearAlgebra
+using QuadGK
+using StaticArrays
 
-println("| Function | Domain | Points | TS (ns) | TS SIMD (ns) | GQ (ns) | Ratio (TS/GQ) | Ratio (TS SIMD/GQ) |")
-println("| :--- | :--- | :--- | :--- | :--- | :--- | :--- | :--- |")
+const RTOL = 1e-6
+const ATOL = 1e-8
+const MAXEVALS = 200_000
+const CUBA_MINEVALS = 2_000
+const BENCH_SAMPLES = 3
+const BENCH_EVALS = 1
 
-const f1 = exp
-const f2 = x -> sin(x)^2
-const f3 = x -> 1.0 / (1.0 + 25.0 * x^2)
-const f4 = x -> sqrt(1.0 - x^2)
-const f5 = x -> x^2
-const f6 = x -> log(1.0 - x)
-const f7 = x -> x^3 # Odd
-const f8 = x -> x^3 + x^2 + x + 1.0 # No symmetry
+const LOW1 = -1.0
+const UP1 = 1.0
+const LOW2 = SVector(-1.0, -1.0)
+const UP2 = SVector(1.0, 1.0)
+const LOW3 = SVector(-1.0, -1.0, -1.0)
+const UP3 = SVector(1.0, 1.0, 1.0)
+const LOW1_VEC = [LOW1]
+const UP1_VEC = [UP1]
+const LOW2_VEC = collect(LOW2)
+const UP2_VEC = collect(UP2)
+const LOW3_VEC = collect(LOW3)
+const UP3_VEC = collect(UP3)
 
-# All tests are on [-1, 1] for now
-const domain = "[-1, 1]"
+const CANDIDATES_AVX_1D = [8, 12, 16, 24, 32, 48, 64, 80, 120, 160, 240, 320, 480, 640]
+const CANDIDATES_AVX_2D = [8, 12, 16, 24, 32, 40, 48, 64, 80, 96, 120, 160]
+const CANDIDATES_AVX_3D = [4, 6, 8, 10, 12, 16, 20, 24, 32, 40, 48, 64]
+const CANDIDATES_GQ_1D = [4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096]
 
-function run_benchmark()
-    functions = [
-        ("exp(x)", f1),
-        ("sin(x)^2", f2),
-        ("1/(1+25x^2)", f3),
-        ("sqrt(1-x^2)", f4),
-        ("x^2", f5),
-        ("log(1-x)", f6),
-        ("x^3", f7),
[…]
