+++ README.md
@@ -242,32 +242,36 @@ println(val)  # ≈ 1.0
 
 ### Results
 
-Comparison against `FastGaussQuadrature.jl` for various functions.
-- `TS`: `integrate1D` (standard Tanh-Sinh)
-- `TS SIMD`: `integrate1D_avx` (SIMD-optimized Tanh-Sinh)
-- `GQ`: `FastGaussQuadrature.jl` (Gauss-Legendre)
+Comparison across:
+- `FastTanhSinhQuadrature.jl` (`adaptive_integrate_*`, `quad`, `_avx`)
+- `QuadGK.jl`
+- `HCubature.jl`
+- `Cubature.jl` (`h` and `p`)
+- `Cuba.jl` (`Vegas`, `Divonne`, `Cuhre`, for >1D)
+- `FastGaussQuadrature.jl` (1D)
 
-Timings are in nanoseconds (ns).
+Methodology:
+- Domain: `[-1,1]^d`
+- Tolerances: `rtol = 1e-6`, `atol = 1e-8`
+- Max evaluations (external adaptive solvers): `200000`
+- Timing: `@belapsed` with interpolation (`samples=3`, `evals=1`)
 
-| Function | Domain | Points | TS (ns) | TS SIMD (ns) | GQ (ns) | Ratio (TS/GQ) | Ratio (TS SIMD/GQ) |
-| :--- | :--- | :--- | :--- | :--- | :--- | :--- | :--- |
-| exp(x) | [-1, 1] | 5 | 33.09 | 14.27 | 33.03 | 1.00 | 0.43 |
-| exp(x) | [-1, 1] | 50 | 322.87 | 81.99 | 176.68 | 1.83 | 0.46 |
-| exp(x) | [-1, 1] | 500 | 3374.75 | 782.08 | 1665.50 | 2.03 | 0.47 |
-| sin(x)^2 | [-1, 1] | 5 | 40.39 | 23.38 | 32.63 | 1.24 | 0.72 |
-| sin(x)^2 | [-1, 1] | 50 | 404.32 | 145.48 | 176.75 | 2.29 | 0.82 |
-| sin(x)^2 | [-1, 1] | 500 | 4361.71 | 1409.10 | 1768.40 | 2.47 | 0.80 |
-| 1/(1+25x^2) | [-1, 1] | 5 | 3.00 | 3.75 | 17.35 | 0.17 | 0.22 |
-| 1/(1+25x^2) | [-1, 1] | 50 | 41.82 | 39.17 | 34.77 | 1.20 | 1.13 |
-| 1/(1+25x^2) | [-1, 1] | 500 | 477.21 | 213.24 | 320.39 | 1.49 | 0.67 |
-| sqrt(1-x^2) | [-1, 1] | 5 | 4.06 | 4.69 | 20.31 | 0.20 | 0.23 |
-| sqrt(1-x^2) | [-1, 1] | 50 | 68.01 | 45.31 | 69.68 | 0.98 | 0.65 |
-| sqrt(1-x^2) | [-1, 1] | 500 | 636.53 | 313.13 | 743.43 | 0.86 | 0.42 |
-| x^2 | [-1, 1] | 5 | 2.01 | 3.04 | 19.86 | 0.10 | 0.15 |
-| x^2 | [-1, 1] | 50 | 17.72 | 21.68 | 38.07 | 0.47 | 0.57 |
-| x^2 | [-1, 1] | 500 | 213.31 | 54.97 | 273.19 | 0.78 | 0.20 |
-| log(1-x) | [-1, 1] | 5 | 38.06 | 30.29 | 34.32 | 1.11 | 0.88 |
-| log(1-x) | [-1, 1] | 50 | 371.88 | 187.29 | 214.70 | 1.73 | 0.87 |
[…]

[…]
