### 2026-02-26 3cbeadc
Bump version to v1.1.0 for next release
