### 2026-02-26 7086925
Fix Test stdlib UUID in Project.toml for Registrator
