+++ Project.toml
@@ -18,7 +18,7 @@ julia = "1.9, 1.10, 1.11, 1.12"
 
 [extras]
 DoubleFloats = "497a8b3b-efae-58df-a0af-a86822472b78"
-Test = "8dfed614-2900-4131-9a72-6804100913c7"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["DoubleFloats", "Test"]
