### 2026-02-26 2db2181
Add 2D integration overload tests and fix default integrate2D
