### 2026-02-28 0d9dd85
ci: upgrade actions and fix codecov integration with v5 and token
