### 2026-02-28 29d458c
ci: add codecov config, version flags, and improve test coverage
