### 2026-02-28 311d895
test: relax brittle tolerances in new CI assertions
