### 2026-02-28 ff6e76a
test: add quad dispatch coverage cases
