### 2026-02-28 5bbc320
test: cover integrate3D_avx vector-bound overload
