### 2026-03-19 47e26f2
Address reviewer benchmark notes in JOSS paper
