+++ JOSS_paper/benchmark_summary_plot.jl
@@ -1,0 +1,143 @@
+using CairoMakie
+
+struct BenchmarkRow
+    dim::String
+    fname::String
+    method::String
+    time_ns::Float64
+    status::String
+end
+
+function read_rows(path::AbstractString)
+    lines = readlines(path)
+    rows = BenchmarkRow[]
+    for line in Iterators.drop(lines, 1)
+        fields = split(chomp(line), ',')
+        length(fields) >= 9 || continue
+        push!(rows, BenchmarkRow(
+            fields[1],
+            fields[2],
+            fields[3],
+            parse(Float64, fields[4]),
+            fields[9],
+        ))
+    end
+    return rows
+end
+
+function benchmark_order(case)
+    order = Dict(
+        ("1D", "1/(1+25x^2)") => 1,
+        ("1D", "x^6 - 2x^3 + 0.5") => 2,
+        ("1D", "log(1-x)") => 3,
+        ("1D", "1/sqrt(1-x^2)") => 4,
+        ("2D", "x^2 + y^2") => 5,
+        ("2D", "exp(x+y)") => 6,
+        ("2D", "1/sqrt((1-x^2)(1-y^2))") => 7,
+        ("3D", "x^2*y^2*z^2") => 8,
+        ("3D", "exp(x+y+z)") => 9,
+        ("3D", "1/sqrt((1-x^2)(1-y^2)(1-z^2))") => 10,
+    )
+    return order[case]
+end
+
+function label_for(case, baseline)
+    dim, fname = case
+    baseline_label = Dict(
+        "FastGauss (precomputed)" => "FastGauss",
+        "QuadGK" => "QuadGK",
+        "CubatureJLh" => "Cubature h",
+        "CubatureJLp" => "Cubature p",
+        "HCubature" => "HCubature",
+        "CubaCuhre" => "Cuba Cuhre",
+    )[baseline]
+    fname_label = Dict(
+        "1/(1+25x^2)" => "1/(1+25x^2)",
+        "x^6 - 2x^3 + 0.5" => "polynomial",
+        "log(1-x)" => "log(1-x)",
+        "1/sqrt(1-x^2)" => "(1-x^2)^(-1/2)",
+        "x^2 + y^2" => "x^2+y^2",
+        "exp(x+y)" => "exp(x+y)",
+        "1/sqrt((1-x^2)(1-y^2))" => "endpoint-singular",
+        "x^2*y^2*z^2" => "x^2 y^2 z^2",
+        "exp(x+y+z)" => "exp(x+y+z)",
+    )[fname]
+    return "$(dim) $(fname_label) [$(baseline_label)]"
+end
+
+function summarize(rows)
+    grouped = Dict{Tuple{String, String}, Vector{BenchmarkRow}}()
+    for row in rows
+        key = (row.dim, row.fname)
+        push!(get!(grouped, key, BenchmarkRow[]), row)
+    end
+
+    cases = sort(collect(keys(grouped)); by=benchmark_order)
+    labels = String[]
+    quad_speedups = Float64[]
+    avx_speedups = Float64[]
+
+    for case in cases
+        case_rows = grouped[case]
+        others = filter(r -> !startswith(r.method, "FTS") && r.status == "ok", case_rows)
+        isempty(others) && continue
+        baseline = findmin(r -> r.time_ns, others)[2]
+        baseline_row = others[baseline]
+
+        quad_row = only(filter(r -> r.method == "FTS quad (convenience)", case_rows))
+        avx_row = only(filter(r -> r.method == "FTS avx (precomputed)", case_rows))
+
+        push!(labels, label_for(case, baseline_row.method))
+        push!(quad_speedups, baseline_row.time_ns / quad_row.time_ns)
[…]
