### 2026-03-20 bdbc5d1
Add benchmark tolerances to JOSS paper
