### 2026-03-20 59aac72
Add future work and contribution guidance to JOSS paper
