### 2026-03-20 0eeeec6
Refine JOSS future work roadmap
