### 2026-03-20 1c97c31
Update README with benchmark plot and contributing section
