### 2026-03-20 b6dd3dd
future work joss paper
