### 2026-03-20 #10076 @svretina
Hi,

I have adressed the two final issues. Should I trigger the pdf compilation here again?
