### 2026-03-21 3f625e4
Fix affiliation formatting for author Stamatis Vretinaris
I added quoting to 1,2,3
