### 2026-03-21 8b21b3b
fix affiliations index order
