### 2026-03-22 #10076 @mongibellili
@svretina see final comments under paper notes issue (closed issue 13)
