### 2026-03-22 7071a5b
Fix missing JOSS paper references
