### 2026-03-22 1910b60
Mention Float64x2 support in JOSS paper
