### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#19 @ranocha
References in the JOSS paper
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

Please check the capitalization of the references. For example, "Johnson, S. G. (2013). QuadGK.jl: Gauss-kronrod integration in julia. https://github.com/128JuliaMath/QuadGK.jl129" should be fixed (Kronrod, Julia).
