### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#20 @ranocha
JOSS Review
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

I have a few questions/comments, where I need your help to perform the JOSS review based on their checklist:

- [ ] **Collaborative effort**: Does the project show evidence of community engagement such as contributions from multiple developers, issues/discussions from external users, or community influence on the development described in the paper?

  I did not see any issues/PRs from other people or similar signs of community engagement. Please tell me more about this aspect so that I can check this box.

- [ ] **Scope and significance**: Does this submission demonstrate clear research impact or credible scholarly significance as described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html)?

  Are there any publications using your software?

- [ ] **Data sharing**: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
  **Reproducibility**: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

  How can I reproduce the benchmark results and figures shown in the paper?
