### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#21 @ranocha
JOSS Paper Figures
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

The figures are hardly readable in the paper. Please increase the font size and use different marker/line types to make the figures readable as well for people affected by colorblindness.
