### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#22 @ranocha
How to use Float32 with quad?
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

The README and documentation mention multiple-precision computations several times. How can I use `Float32`? Based on examples like https://svretina.github.io/FastTanhSinhQuadrature.jl/stable/examples/basics/#3.-High-Precision-Integration-(Double64,-BigFloat), I would have expected to get `Float32` from the last example below:

```julia
julia> quad(exp, 0.0, 1.0) |> typeof
Float64

julia> quad(exp, big(1.0), big(1.0)) |> typeof
BigFloat

julia> quad(exp, 0.0f0, 1.0f0) |> typeof
Float64
```
