### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#23 @ranocha
Documentation
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

- [ ] https://svretina.github.io/FastTanhSinhQuadrature.jl/stable/theory/#Overview: $$w(t)$$ in the last equation of this section is not defined previously.
- [ ] The figure in https://svretina.github.io/FastTanhSinhQuadrature.jl/stable/theory/#Transformation-Visualization does not load on my computer.
- [ ] How shall one estimate the number of nodes required for `integrate1D_avx` to get the desired accuracy (obtained via a tolerance in `quad`)?
- [ ]  A lot of content from https://svretina.github.io/FastTanhSinhQuadrature.jl/stable/examples/basics/ and https://svretina.github.io/FastTanhSinhQuadrature.jl/stable/examples/advanced/ seems to be duplicated. For example, just looking at the titles, both pages have "pre-computing nodes", "SIMD", "singularities", and "arbitrary precision".
- [ ] The results shown on https://svretina.github.io/FastTanhSinhQuadrature.jl/stable/benchmarks/#Results are not what I would be looking for when trying to decide whether to use this package: I would not be interested in knowing how long it takes to compute an approximation for a given number of points, but for a given accuracy of the result.
- [ ] Tolerance of `quad`: What kind of tolerance is `tol` in https://svretina.github.io/FastTanhSinhQuadrature.jl/stable/api/#FastTanhSinhQuadrature.quad? Typically, a combination of absolute and relative tolerance like in https://juliamath.github.io/QuadGK.jl/stable/api/#QuadGK.quadgk is recommended.
- [ ] Can you please add this takeaway message summarizing when to use which package to a section such as "Other Julia quadrature packages" in the README.md and your documentation, similar to https://juliamath.github.io/QuadGK.jl/stable/#Other-Julia-quadrature-packages?
