### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#24 @ranocha
Unnecessary restriction to `Function`
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

In many cases, it can be helpful to use `struct`s as custom functors, e.g., to carry parameters around. In this case, the restriction to `Function` is a disadvantage, e.g.,

```julia
julia> using FastTanhSinhQuadrature

julia> struct Functor{C}
           c::C
       end

julia> f = Functor(10.0)
Functor{Float64}(10.0)

julia> quad(f, 0.0, 1.0)
ERROR: MethodError: no method matching quad(::Functor{Float64}, ::Float64, ::Float64)

Closest candidates are:
  quad(::Function, ::T, ::T; tol, max_levels) where T<:Real
   @ FastTanhSinhQuadrature ~/.julia/packages/FastTanhSinhQuadrature/T5vJb/src/quad.jl:31
  quad(::Function, ::Real, ::Real; tol, max_levels)
   @ FastTanhSinhQuadrature ~/.julia/packages/FastTanhSinhQuadrature/T5vJb/src/quad.jl:47

Stacktrace:
 [1] top-level scope
   @ REPL[4]:1
```
