### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#25 @ranocha
Comparison with other packages and reliability of the error estimate
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

When playing around with this package, I noticed the following behavior:

```julia
julia> using FastTanhSinhQuadrature, QuadGK

julia> k = 10^3; quadgk(-pi, pi) do x
           sin(k * x)^2
       end |> x -> first(x) - pi
2.6645352591003757e-15

julia> k = 10^3; quad(-pi, pi) do x
           sin(k * x)^2
       end - pi
0.08430370475410198
```

Please comment on the error estimator and the general behavior of this package compared to alternatives like QuadGK.jl, which appear to handle a case like this much more robustly.
