### 2026-03-23 svretina/FastTanhSinhQuadrature.jl#26 @ranocha
Wrong research impact statement in paper
Thank you for your contribution to the open-source ecosystem! This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/10076.

> FastTanhSinhQuadrature.jl has been integrated as a backend for Integrals.jl

This is not true as far as I can tell. There is an open PR https://github.com/SciML/Integrals.jl/pull/310 created by a bot without visible human interaction, but the integration has not been finished yet as far as I know.
