### 2026-03-23 #10076 @svretina
I will adress the issues raised and generate the a new pdf here, once I am finished.
