### 2026-03-24 #10076 @ranocha
Hi @vissarion! I have two questions on the review of this submission.

#### Question 1

Concerning the **Collaborative effort** item on the checklist, the author wrote in https://github.com/svretina/FastTanhSinhQuadrature.jl/issues/20#issuecomment-4111588206:

> this project started from my own research needs and I decided to put some effort to make it more accessible to the community since I found that the Julia ecosystem was lacking a tanh-sinh quadrature library which was focused on performance. So this is entirely my personal effort, although contribution and collaboration is encouraged.

I can obviously not check the box, since there is no evidence of community engagement. How does JOSS handle such a case?

#### Question 2

Concerning the **Scope and significance** item on the checklist, the author wrote in https://github.com/svretina/FastTanhSinhQuadrature.jl/issues/20#issuecomment-4111588206:

> Not yet, the research I used it is at the stage of writing the draft and I will submit in the next months. So there is no publication using this package yet.

Thus, there is no evidence of publications or analyses using the software. How does JOSS handle such a case?
