### 2026-03-30 #10076 @danielskatz
Hi - As the track editor, I wanted to say that we are discussing this. We changed our scope requirements at the start of the year, and we're still in a period of determining practices for best interpreting and implementing them, and ensuring that the language in our docs and checklists match them. Since this submission came in and we evaluated its scope, we have made some small changes in the checklists, and we now need to determine how to deal with this inconsistency.  We'll get back to you shortly.
