### 2026-03-30 #10076 @vissarion
Hi, we're going to let this proceed without these two items (Collaborative effort & Scope and significance) on the checklist being checked off, for the reason @danielskatz mentioned above.

@ranocha thanks for raising this issue, could you please continue on the rest of the review without these two items?
