### 2026-04-01 #10076 @vissarion
> I created several issues in the repository that are still open.

Great, thanks! @svretina could you please address the issues raised by @ranocha ?
