### 2026-04-02 c37f651
Fix bibliography title capitalization
