+++ JOSS_paper/paper.bib
@@ -21,7 +21,7 @@
 
 @article{Virtanen2020,
   author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and {van der Walt}, St{\'e}fan J. and Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C J and Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and Harris, Charles R. and Archibald, Anne M. and Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in Python}},
+  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in {Python}}},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -32,42 +32,42 @@
 
 @misc{Boost,
   author = {{Boost C++ Libraries}},
-  title  = {Boost C++ Libraries},
+  title  = {{Boost {C++} Libraries}},
   url    = {https://www.boost.org/},
   year   = {2024}
 }
 
 @misc{Johansson_mpmath,
   author = {Johansson, Fredrik and others},
-  title  = {mpmath: a {Python} library for arbitrary-precision floating-point arithmetic},
+  title  = {{mpmath}: a {Python} library for arbitrary-precision floating-point arithmetic},
   url    = {http://mpmath.org/},
   year   = {2023}
 }
 
 @misc{QuadGK,
-  author = {Johnson, Steven G.},
-  title  = {QuadGK.jl: Gauss-Kronrod integration in Julia},
-  url    = {https://github.com/JuliaMath/QuadGK.jl},
-  year   = {2013}
+  title = {{QuadGK.jl}: {G}auss--{K}ronrod integration in {J}ulia},
+  author = {Steven G. Johnson},
+  year = {2013},
+  howpublished = {\url{https://github.com/JuliaMath/QuadGK.jl}}
 }
 
 @misc{HCubature,
-  author = {Johnson, Steven G.},
-  title  = {HCubature.jl: H-adaptive multidimensional integration in Julia},
-  url    = {https://github.com/JuliaMath/HCubature.jl},
-  year   = {2017}
+  title = {The {HCubature.jl} package for multi-dimensional adaptive integration in {Julia}},
+  author = {Steven G. Johnson},
+  year = {2017},
+  howpublished = {\url{https://github.com/JuliaMath/HCubature.jl}}
 }
 
 @misc{Cubature,
-  author = {Johnson, Steven G.},
-  title  = {Multi-dimensional adaptive integration in {C}: The {Cubature} package},
-  url    = {https://github.com/stevengj/cubature},
-  year   = {2005}
+  title = {Multi-dimensional adaptive integration in {C}: The {Cubature} package},
+  author = {Steven G. Johnson},
+  year = {2005},
+  howpublished = {\url{https://github.com/stevengj/cubature}}
 }
 
 @article{Cuba,
   author  = {Hahn, Thomas},
-  title   = {Cuba: A library for multidimensional numerical integration},
+  title   = {{Cuba}: A library for multidimensional numerical integration},
[…]
