### 2026-04-02 0f36c21
Improve JOSS figure readability
