### 2026-04-02 2018408
Fix type stability across precision types
