+++ test/TypeStabilityTests.jl
@@ -16,12 +16,16 @@
 
 using FastTanhSinhQuadrature
 using JET
+using MultiFloats
+using Quadmath
 using Test
 using StaticArrays
 
 # Test Inputs (Global Scope)
 const T_val = Float64
 const x_vec, w_vec, h_val = tanhsinh(T_val, Val(10))
+const T32_val = Float32
+const x32_vec, w32_vec, h32_val = tanhsinh(T32_val, Val(10))
 
 # 1D Bounds
 const low1d_val, up1d_val = -1.0, 1.0
@@ -31,11 +35,16 @@ const low2d_val = SVector(-1.0, -1.0)
 const up2d_val = SVector(1.0, 1.0)
 const low3d_val = SVector(-1.0, -1.0, -1.0)
 const up3d_val = SVector(1.0, 1.0, 1.0)
+const low2d32_val = SVector(-1.0f0, -1.0f0)
+const up2d32_val = SVector(1.0f0, 1.0f0)
+const low3d32_val = SVector(-1.0f0, -1.0f0, -1.0f0)
+const up3d32_val = SVector(1.0f0, 1.0f0, 1.0f0)
 
 # Simple polynomial functions
 func1(x) = x^2
 func2(x, y) = x * y
 func3(x, y, z) = x * y * z
+func_cmpl(x, bmx, xma) = inv(sqrt(bmx * xma))
 
 @testset "Type Stability" begin
     # JET 0.9 uses `target_defined_modules`, while JET 0.11 uses `target_modules`.
@@ -62,62 +71,194 @@ func3(x, y, z) = x * y * z
     check_opt(tanhsinh, (Type{Float64}, Val{10}))
     check_call(tanhsinh, (Type{Float64}, Val{10}))
 
+    check_opt(tanhsinh, (Type{Float32x2}, Int))
+    check_call(tanhsinh, (Type{Float32x2}, Int))
+
+    check_opt(tanhsinh, (Type{Float64x2}, Int))
+    check_call(tanhsinh, (Type{Float64x2}, Int))
+
+    check_opt(tanhsinh, (Type{BigFloat}, Int))
+    check_call(tanhsinh, (Type{BigFloat}, Int))
+
+    check_opt(tanhsinh, (Type{Double64}, Int))
+    check_call(tanhsinh, (Type{Double64}, Int))
+
+    check_opt(tanhsinh, (Type{Float128}, Int))
+    check_call(tanhsinh, (Type{Float128}, Int))
+
     # --- Integrate 1D ---
     check_opt(integrate1D, (typeof(func1), Vector{Float64}, Vector{Float64}, Float64))
     check_call(integrate1D, (typeof(func1), Vector{Float64}, Vector{Float64}, Float64))
 
+    check_opt(integrate1D, (typeof(func1), Vector{Float32}, Vector{Float32}, Float32))
+    check_call(integrate1D, (typeof(func1), Vector{Float32}, Vector{Float32}, Float32))
+
     check_opt(integrate1D, (typeof(func1), Float64, Float64, Vector{Float64}, Vector{Float64}, Float64))
     check_call(integrate1D, (typeof(func1), Float64, Float64, Vector{Float64}, Vector{Float64}, Float64))
+
+    check_opt(integrate1D, (typeof(func1), Float32, Float32, Vector{Float32}, Vector{Float32}, Float32))
+    check_call(integrate1D, (typeof(func1), Float32, Float32, Vector{Float32}, Vector{Float32}, Float32))
+
+    check_opt(integrate1D, (typeof(func1), Float32x2, Float32x2, Vector{Float32x2}, Vector{Float32x2}, Float32x2))
+    check_call(integrate1D, (typeof(func1), Float32x2, Float32x2, Vector{Float32x2}, Vector{Float32x2}, Float32x2))
+
[…]
