### 2026-04-02 35e7403
Keep adaptive cache counters as Int
