### 2026-04-02 9ffa643
Allow callable functors in quadrature API
