### 2026-04-03 91e92de
Improve adaptive quadrature convergence checks
