+++ test/runtests.jl
@@ -205,12 +205,12 @@ end
         val_true = exp(1.0) - exp(0.0)
 
         # Using quad which uses adaptive_integrate_1D
-        val = quad(f, 0.0, 1.0; tol=1e-8)
+        val = quad(f, 0.0, 1.0; rtol=1e-8)
         @test isapprox(val, val_true, atol=1e-8)
 
         # Test with higher precision
         if Double64 in Types
-            val_d = quad(f, Double64(0.0), Double64(1.0); tol=1e-15)
+            val_d = quad(f, Double64(0.0), Double64(1.0); rtol=1e-15)
             @test isapprox(val_d, val_true, atol=1e-15)
         end
     end
@@ -239,30 +239,30 @@ end
         @test inferred_return_type(() -> integrate1D(f1, 0.0f0, 1.0f0, x32, w32, h32)) === Float32
         @test inferred_return_type(() -> integrate1D_avx(f1, x32, w32, h32)) === Float32
         @test inferred_return_type(() -> integrate1D_avx(f1, 0.0f0, 1.0f0, x32, w32, h32)) === Float32
-        @test inferred_return_type(() -> adaptive_integrate_1D(Float32, f1, 0.0f0, 1.0f0; tol=1f-5, max_levels=0)) === Float32
-        @test inferred_return_type(() -> adaptive_integrate_1D_cmpl(Float32, f_cmpl, -1.0f0, 1.0f0; tol=1f-5, max_levels=0)) === Float32
+        @test inferred_return_type(() -> adaptive_integrate_1D(Float32, f1, 0.0f0, 1.0f0; rtol=1f-5, max_levels=0)) === Float32
+        @test inferred_return_type(() -> adaptive_integrate_1D_cmpl(Float32, f_cmpl, -1.0f0, 1.0f0; rtol=1f-5, max_levels=0)) === Float32
         @test inferred_return_type(() -> quad(f1, 0.0f0, 1.0f0; max_levels=0)) === Float32
-        @test inferred_return_type(() -> quad(f1, 0.0f0, 1.0f0; tol=1f-5, max_levels=0)) === Float32
+        @test inferred_return_type(() -> quad(f1, 0.0f0, 1.0f0; rtol=1f-5, max_levels=0)) === Float32
         @test inferred_return_type(() -> quad_cmpl(f_cmpl, -1.0f0, 1.0f0; max_levels=0)) === Float32
-        @test inferred_return_type(() -> quad_cmpl(f_cmpl, -1.0f0, 1.0f0; tol=1f-5, max_levels=0)) === Float32
+        @test inferred_return_type(() -> quad_cmpl(f_cmpl, -1.0f0, 1.0f0; rtol=1f-5, max_levels=0)) === Float32
         @test inferred_return_type(() -> quad_split(f_split, 0.0f0, -1.0f0, 1.0f0; max_levels=0)) === Float32
-        @test inferred_return_type(() -> quad_split(f_split, 0.0f0, -1.0f0, 1.0f0; tol=1f-4, max_levels=0)) === Float32
-        @test inferred_return_type(() -> quad_split(f_split, 0.0f0; tol=1f-4, max_levels=0)) === Float32
+        @test inferred_return_type(() -> quad_split(f_split, 0.0f0, -1.0f0, 1.0f0; rtol=1f-4, max_levels=0)) === Float32
+        @test inferred_return_type(() -> quad_split(f_split, 0.0f0; rtol=1f-4, max_levels=0)) === Float32
         @test typed_call_return_type(quad, Tuple{typeof(functor1), Float32, Float32}) === Float32
-        @test inferred_return_type(() -> quad(functor1, 0.0f0, 1.0f0; tol=1f-5, max_levels=0)) === Float32
[…]
