### 2026-04-03 svretina/FastTanhSinhQuadrature.jl#26 @svretina
What would you suggest for this issue?

I would prefer to remove the statement altogether, but the JOSS guidelines are that it needs to be present as a section.
I could rephrase that there are pending plans to integrate in into the `Integrals.jl` package as a backend.
