### 2026-04-04 ee1ce64
Update cache-aware API docs, benchmark methodology, and summary plots
