+++ src/adaptive.jl
@@ -16,24 +16,33 @@
 
 # Adaptive integration functions
 
+@inline function _require_cache_levels(cache, max_levels::Int)
+    length(cache.xs) >= max_levels ||
+        throw(ArgumentError("provided adaptive cache supports $(length(cache.xs)) levels, but `max_levels=$(max_levels)` was requested."))
+    return cache
+end
+
 """
-    adaptive_integrate_1D(::Type{T}, f, a, b; rtol, atol, max_levels::Int=16, warn::Bool=true)
+    adaptive_integrate_1D(::Type{T}, f, a, b; rtol, atol, max_levels::Int=16, warn::Bool=true, cache=nothing)
 
 Adaptive 1D Tanh-Sinh integration over `[a, b]`. Starts with a coarse grid (h ≈ tmax/2) and halves 
 the step size at each level. Reuses function evaluations from previous levels by only computing
 new (odd-indexed) nodes. Exploits symmetry around the center of the interval.
+The error estimate is `abs(I_new - I_old)` between successive refinement levels,
+and refinement stops when it is below `max(atol, rtol * abs(I_new))`.
 """
 function adaptive_integrate_1D(::Type{T}, f::F, a, b;
     rtol=nothing, atol::Real=0, max_levels::Int=16,
-    warn::Bool=true) where {T<:Real,F}
+    warn::Bool=true, cache=nothing) where {T<:Real,F}
     a_T, b_T = T(a), T(b)
     rtol_T, atol_T = _resolve_tolerances(T; rtol=rtol, atol=atol)
     Δx, x₀ = _midpoint_radius(a_T, b_T)
-    cache = _adaptive_1d_cache(T, max_levels)
+    cache1d = cache === nothing ? adaptive_cache_1D(T; max_levels=max_levels) :
+              _require_cache_levels(cache, max_levels)
     half = _half(T)
 
     # Initial Grid (Level 0)
-    h = cache.tm * half
+    h = cache1d.tm * half
 
     # Weight at t=0 is π/2
     w0 = T(π) * half
@@ -41,9 +50,9 @@ function adaptive_integrate_1D(::Type{T}, f::F, a, b;
 
     # Initial Level evaluations
     @inbounds for i in 1:2
-        xk = cache.initial_x[i]
+        xk = cache1d.initial_x[i]
         Δxxk = Δx * xk
-        s_total += cache.initial_w[i] * (f(x₀ + Δxxk) + f(x₀ - Δxxk))
+        s_total += cache1d.initial_w[i] * (f(x₀ + Δxxk) + f(x₀ - Δxxk))
     end
 
     old_res = Δx * h * s_total
@@ -52,8 +61,8 @@ function adaptive_integrate_1D(::Type{T}, f::F, a, b;
     for level in 1:max_levels
         h *= half
         s_new = zero(T)
-        x_level = cache.xs[level]
-        w_level = cache.ws[level]
+        x_level = cache1d.xs[level]
+        w_level = cache1d.ws[level]
         @inbounds for i in eachindex(x_level)
             xk = x_level[i]
             Δxxk = Δx * xk
@@ -76,21 +85,24 @@ function adaptive_integrate_1D(::Type{T}, f::F, a, b;
 end
 
 """
-    adaptive_integrate_2D(::Type{T}, f, low::SVector{2,T}, up::SVector{2,T}; rtol, atol, max_levels::Int=8, warn::Bool=true)
+    adaptive_integrate_2D(::Type{T}, f, low::SVector{2,T}, up::SVector{2,T}; rtol, atol, max_levels::Int=8, warn::Bool=true, cache=nothing)
 
[…]
