+++ JOSS_paper/benchmark_summary_plot.jl
@@ -1,7 +1,9 @@
 using CairoMakie
+using LaTeXStrings
 
 const QUAD_COLOR = RGBAf(0.0, 0.447, 0.698, 1.0)
 const AVX_COLOR = RGBAf(0.835, 0.369, 0.0, 1.0)
+const LINK_COLOR = RGBAf(0.55, 0.55, 0.55, 0.7)
 
 struct BenchmarkRow
     dim::String
@@ -9,6 +11,14 @@ struct BenchmarkRow
     method::String
     time_ns::Float64
     status::String
+end
+
+struct CaseSummary
+    dim::String
+    fname::String
+    baseline::String
+    quad_speedup::Float64
+    avx_speedup::Float64
 end
 
 function read_rows(path::AbstractString)
@@ -32,40 +42,50 @@ function benchmark_order(case)
     order = Dict(
         ("1D", "1/(1+25x^2)") => 1,
         ("1D", "x^6 - 2x^3 + 0.5") => 2,
-        ("1D", "log(1-x)") => 3,
-        ("1D", "1/sqrt(1-x^2)") => 4,
-        ("2D", "x^2 + y^2") => 5,
-        ("2D", "exp(x+y)") => 6,
-        ("2D", "1/sqrt((1-x^2)(1-y^2))") => 7,
-        ("3D", "x^2*y^2*z^2") => 8,
-        ("3D", "exp(x+y+z)") => 9,
-        ("3D", "1/sqrt((1-x^2)(1-y^2)(1-z^2))") => 10,
+        ("1D", "sin^2(1000x)") => 3,
+        ("1D", "log(1-x)") => 4,
+        ("1D", "1/sqrt(1-x^2)") => 5,
+        ("2D", "x^2 + y^2") => 6,
+        ("2D", "exp(x+y)") => 7,
+        ("2D", "1/sqrt((1-x^2)(1-y^2))") => 8,
+        ("3D", "x^2*y^2*z^2") => 9,
+        ("3D", "exp(x+y+z)") => 10,
+        ("3D", "1/sqrt((1-x^2)(1-y^2)(1-z^2))") => 11,
     )
     return order[case]
 end
 
-function label_for(case, baseline)
-    dim, fname = case
-    baseline_label = Dict(
-        "FastGauss (precomputed)" => "FastGauss",
-        "QuadGK" => "QuadGK",
-        "CubatureJLh" => "Cubature h",
-        "CubatureJLp" => "Cubature p",
-        "HCubature" => "HCubature",
-        "CubaCuhre" => "Cuba Cuhre",
-    )[baseline]
-    fname_label = Dict(
-        "1/(1+25x^2)" => "1/(1+25x^2)",
-        "x^6 - 2x^3 + 0.5" => "polynomial",
-        "log(1-x)" => "log(1-x)",
-        "1/sqrt(1-x^2)" => "(1-x^2)^(-1/2)",
-        "x^2 + y^2" => "x^2+y^2",
-        "exp(x+y)" => "exp(x+y)",
-        "1/sqrt((1-x^2)(1-y^2))" => "endpoint-singular",
-        "x^2*y^2*z^2" => "x^2 y^2 z^2",
-        "exp(x+y+z)" => "exp(x+y+z)",
+function baseline_tex(method::String)
+    return Dict(
+        "FastGauss (precomputed)" => raw"\mathrm{FastGauss}",
+        "QuadGK" => raw"\mathrm{QuadGK}",
+        "CubatureJLh" => raw"\mathrm{Cubature\ h}",
+        "CubatureJLp" => raw"\mathrm{Cubature\ p}",
[…]
