### 2026-04-04 4c831a6
Fix JET inference in adaptive cache builders
