+++ docs/src/benchmarks.md
@@ -18,6 +18,9 @@ Cross-library comparisons include:
 - For each test case, we infer an `N` from the level where `FTS adaptive`
   satisfies the stopping criterion.
 - `FTS avx` is benchmarked at that adaptive-matched `N`.
+- In the summary figure, the SIMD path is shown for both newer Julia
+  (where `@turbo` can degrade to a fallback) and Julia 1.12
+  (where `LoopVectorization` is active), so the version effect is visible.
 - `FastGauss` (1D) is calibrated independently to the same tolerance target.
 - The table therefore reports adaptive tolerance-based runs plus fixed-grid
   comparisons where each method uses its own convergence path.
@@ -36,6 +39,10 @@ is usually the problem class:
 
 - Domain: case-dependent (most tests use `[-1,1]^d`; oscillatory 1D test uses `[-π,π]`)
 - Tolerances: `rtol = 1e-6`, `atol = 1e-8`
+- Same target is used across libraries:
+  `quad`/`adaptive_integrate_*`/`QuadGK`/`HCubature`/`Cuba` use `rtol`+`atol`,
+  `Cubature.jl` uses `reltol`+`abstol` set to the same values,
+  and `FastGauss` is calibrated to satisfy the same target in terms of actual error.
 - Max evaluations (external adaptive solvers): `200000`
 - Timing: `@belapsed` with interpolation (`samples=3`, `evals=1`)
 - One warm call is executed before each timed benchmark.
@@ -57,17 +64,17 @@ Raw output files are generated at:
 
 | Dim | Function | FTS adaptive | FTS quad | FTS avx | QuadGK | HCubature | Cubature h | Cubature p | Cuba Vegas | Cuba Divonne | Cuba Cuhre | FastGauss |
 | :-- | :------- | ----------: | -------: | ------: | -----: | --------: | ---------: | ---------: | ---------: | -----------: | ---------: | --------: |
-| 1D | 1/(1+25x^2) | 735.0000 | 628.0000 | 474.0000 | 1612.0000 | 1.620e+04 | 1.644e+04 | 2.007e+04 | n/a | n/a | n/a | 176.0000 |
-| 1D | 1/sqrt(1-x^2) | 574.0000 | 528.0000 | 523.0000 | 1.154e+04 | 2.189e+05 | 2.389e+05 | 2834.0000 * | n/a | n/a | n/a | 1.146e+04 * |
-| 1D | log(1-x) | 1013.0000 | 592.0000 | 442.0000 | 5237.0000 | 6.031e+04 | 7.402e+04 | 1356.0000 * | n/a | n/a | n/a | 1.208e+04 |
-| 1D | sin^2(1000x) | 1.770e+05 | 2.542e+05 | 2.909e+04 | 1.142e+04 | 1.313e+05 | 1.041e+05 | 1301.0000 * | n/a | n/a | n/a | 3.748e+04 |
-| 1D | x^6 - 2x^3 + 0.5 | 1306.0000 | 803.0000 | 335.0000 | 1898.0000 | 4813.0000 | 2249.0000 | 8949.0000 | n/a | n/a | n/a | 118.0000 |
[…]
