### 2026-04-04 8cfd19c
Update JOSS draft with reproducibility/Zenodo and refreshed artifacts
