+++ JOSS_paper/paper.bib
@@ -117,3 +117,12 @@ abstract = {Parallel features of the multidimensional numerical integration pack
  volume = {5},
  year = {2017}
 }
+
+@misc{FastTanhSinhZenodo2026,
+  author       = {Vretinaris, Stamatis},
+  title        = {{FastTanhSinhQuadrature.jl}},
+  year         = {2026},
+  doi          = {10.5281/zenodo.19420466},
+  url          = {https://zenodo.org/records/19420466},
+  howpublished = {\url{https://zenodo.org/records/19420466}}
+}
