+++ benchmark/results/timings_summary_julia1.12.md
@@ -2,17 +2,17 @@
 
 | Dim | Function | FTS adaptive | FTS quad | FTS avx | QuadGK | HCubature | Cubature h | Cubature p | Cuba Vegas | Cuba Divonne | Cuba Cuhre | FastGauss |
 | :-- | :------- | ----------: | -------: | ------: | -----: | --------: | ---------: | ---------: | ---------: | -----------: | ---------: | --------: |
-| 1D | 1/(1+25x^2) | 715.0000 | 560.0000 | 331.0000 | 522.0000 | 1.325e+04 | 1.212e+04 | 1.514e+04 | n/a | n/a | n/a | 130.0000 |
-| 1D | 1/sqrt(1-x^2) | 468.0000 | 401.0000 | 312.0000 | 7608.0000 | 2.109e+05 | 2.979e+05 | 2191.0000 * | n/a | n/a | n/a | 1.300e+04 * |
-| 1D | log(1-x) | 727.0000 | 647.0000 | 379.0000 | 5125.0000 | 5.575e+04 | 8.926e+04 | 1507.0000 * | n/a | n/a | n/a | 1.337e+04 |
-| 1D | sin^2(1000x) | 1.936e+05 | 1.840e+05 | 3.772e+04 | 1.363e+04 | 1.511e+05 | 1.003e+05 | 1204.0000 * | n/a | n/a | n/a | 3.140e+04 |
-| 1D | x^6 - 2x^3 + 0.5 | 1491.0000 | 584.0000 | 283.0000 | 418.0000 | 7145.0000 | 3120.0000 | 3697.0000 | n/a | n/a | n/a | 113.0000 |
-| 2D | 1/sqrt((1-x^2)(1-y^2)) | 3308.0000 | 2934.0000 | 1781.0000 | n/a | 6.426e+07 | 3.079e+07 | 1905.0000 * | 1.016e+08 * | 1.492e+08 * | 7.848e+07 * | n/a |
-| 2D | exp(x+y) | 1.921e+04 | 1.736e+04 | 5146.0000 | n/a | 8.382e+04 | 4.263e+04 | 4.012e+04 | 9.871e+07 * | 9.276e+07 | 6.636e+04 | n/a |
-| 2D | x^2 + y^2 | 3403.0000 | 1981.0000 | 921.0000 | n/a | 8331.0000 | 2865.0000 | 4913.0000 | 1.218e+08 * | 1.010e+08 | 6.841e+04 | n/a |
-| 3D | 1/sqrt((1-x^2)(1-y^2)(1-z^2)) | 9.817e+04 | 2.139e+05 | 6.309e+04 | n/a | 3.905e+07 * | 3.127e+07 * | 4478.0000 * | 1.272e+08 * | 1.460e+08 * | 1.267e+08 * | n/a |
-| 3D | exp(x+y+z) | 1.132e+06 | 1.078e+06 | 3.323e+05 | n/a | 3.120e+05 | 3.067e+05 | 8.971e+05 | 1.312e+08 * | 1.512e+08 * | 1.532e+05 | n/a |
-| 3D | x^2*y^2*z^2 | 6.353e+04 | 6.316e+04 | 1.960e+04 | n/a | 1.241e+07 | 1.045e+07 | 2.553e+04 | 1.349e+08 * | 1.540e+08 * | 4.159e+07 | n/a |
+| 1D | 1/(1+25x^2) | 640.0000 | 477.0000 | 370.0000 | 679.0000 | 1.237e+04 | 1.873e+04 | 1.369e+04 | n/a | n/a | n/a | 114.0000 |
[…]

[…]
