### 2026-04-04 #10076 @svretina
Dear Editors and Reviewers,

Thank you for the detailed feedback throughout the review process. I have now addressed all issues raised in the reviewer comments.

In this update, I:
- implemented the requested code and API fixes,
- updated tests and benchmarks accordingly,
- revised the documentation and JOSS draft for consistency and clarity,
- added explicit reproducibility instructions (`REPRODUCIBILITY.md`),
- released a new software version,
- and archived the exact release snapshot on Zenodo for long-term reproducibility and citation.

Zenodo archive (reproducibility snapshot): https://zenodo.org/records/19420466
DOI: https://doi.org/10.5281/zenodo.19420466
