### 2026-04-05 d806d9d
Optimize integration hot paths
