+++ src/caches.jl
@@ -1,0 +1,166 @@
+# FastTanhSinhQuadrature.jl: Fast and high-precision numerical integration using Tanh-Sinh (Double Exponential) quadrature.
+# Copyright (C) 2024-2026 Stamatis Vretinaris
+# 
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+# 
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+# 
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+# Adaptive cache construction and validation
+
+@inline function _require_cache_levels(cache, max_levels::Int)
+    length(cache.xs) >= max_levels ||
+        throw(ArgumentError("provided adaptive cache supports $(length(cache.xs)) levels, but `max_levels=$(max_levels)` was requested."))
+    return cache
+end
+
+struct _Adaptive1DCache{T}
+    tm::T
+    initial_x::NTuple{2,T}
+    initial_w::NTuple{2,T}
+    initial_c::NTuple{2,T}
+    xs::Vector{Vector{T}}
+    ws::Vector{Vector{T}}
+    cs::Vector{Vector{T}}
+end
+
+@inline _adaptive_1d_window(::Type{T}, kind::Symbol) where {T<:Real} =
+    kind === :complement ? t_w_max(T, 1) : tmax(T)
+
+const _ADAPTIVE_1D_CACHES = Dict{Tuple{DataType, Int, Symbol}, Any}()
+const _ADAPTIVE_1D_CACHES_LOCK = ReentrantLock()
+
+function _build_adaptive_1d_cache(::Type{T}, max_levels::Int, kind::Symbol) where {T<:Real}
+    max_levels >= 0 || throw(ArgumentError("`max_levels` must be nonnegative."))
+    tm = _adaptive_1d_window(T, kind)
+    half = _half(T)
+    h = tm * half
+
+    h1 = h
+    h2 = h + h
+    initial_x = (ordinate(h1), ordinate(h2))
+    initial_w = (weight(h1), weight(h2))
+    initial_c = (ordinate_complement(h1), ordinate_complement(h2))
+
+    xs = Vector{Vector{T}}(undef, max_levels)
+    ws = Vector{Vector{T}}(undef, max_levels)
+    cs = Vector{Vector{T}}(undef, max_levels)
+    for level in 1:max_levels
+        h *= half
+        n = 1 << level
+        x = Vector{T}(undef, n)
+        w = Vector{T}(undef, n)
+        c = Vector{T}(undef, n)
+        @inbounds for i in 1:n
+            tk = (2i - 1) * h
+            x[i] = ordinate(tk)
+            w[i] = weight(tk)
+            c[i] = ordinate_complement(tk)
+        end
+        xs[level] = x
+        ws[level] = w
+        cs[level] = c
+    end
+    return _Adaptive1DCache{T}(tm, initial_x, initial_w, initial_c, xs, ws, cs)
+end
+
+function _adaptive_1d_cache(::Type{T}, max_levels::Int, kind::Symbol=:regular) where {T<:Real}
+    key = (T, max_levels, kind)
[…]
