### 2026-04-05 a7d789a
Publish v2.0.0 stable docs
