### 2026-04-05 d140f2b
Refine adaptive kernels and expand coverage
