### 2026-04-05 #10076 @ranocha
Thank you for your contribution to the open-source ecosystem, @svretina, and for improving your package and paper.

@vissarion I have finished my review so far. The only remaining items on the checklist are still open due to the discussion above (https://github.com/openjournals/joss-reviews/issues/10076#issuecomment-4155089125).
