### 2026-04-06 #10076 @vissarion
Thanks @ranocha!

@svretina When a submission is ready to be accepted, we ask that the authors issue a new tagged release of the software (if changed), and archive it (see [this guide](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content)). Please do this and post the version number and archive DOI here.
Please make sure that the author names, affiliations as well as the title of the archive (e.g. zenodo) is exactly the same with the submission. For example, in the zenodo link above please update the title and the affiliations to match the ones in the paper.
