### 2026-04-06 668c6fd
Add adaptive AVX integration paths
