+++ src/FastTanhSinhQuadrature.jl
@@ -46,6 +46,8 @@ export tanhsinh, quad, quad_split, quad_cmpl,
     integrate1D_avx, integrate2D_avx, integrate3D_avx,
     adaptive_integrate_1D, adaptive_integrate_2D, adaptive_integrate_3D,
     adaptive_integrate_1D_cmpl,
+    adaptive_integrate_1D_avx, adaptive_integrate_2D_avx,
+    adaptive_integrate_3D_avx, adaptive_integrate_1D_cmpl_avx,
     adaptive_cache_1D, adaptive_cache_2D, adaptive_cache_3D
 
 end
