+++ src/adaptive.jl
@@ -41,8 +41,7 @@ function adaptive_integrate_1D(::Type{T}, f::F, a_T::T, b_T::T,
     # Initial Level evaluations
     @inbounds for i in 1:2
         xk = cache1d.initial_x[i]
-        Δxxk = Δx * xk
-        s_total += cache1d.initial_w[i] * (f(x₀ + Δxxk) + f(x₀ - Δxxk))
+        s_total += cache1d.initial_w[i] * (f(Δx * xk + x₀) + f(x₀ - Δx * xk))
     end
 
     old_res = Δx * h * s_total
@@ -56,7 +55,7 @@ function adaptive_integrate_1D(::Type{T}, f::F, a_T::T, b_T::T,
         @inbounds for i in eachindex(x_level)
             xk = x_level[i]
             xp = Δx * xk + x₀
-            xm = -Δx * xk + x₀
+            xm = x₀ - Δx * xk
             s_new += w_level[i] * (f(xp) + f(xm))
         end
 
@@ -86,6 +85,64 @@ function adaptive_integrate_1D(::Type{T}, f::F, a, b;
 end
 
 """
+    adaptive_integrate_1D_avx(::Type{T}, f, a, b; rtol, atol, max_levels::Int=16, warn::Bool=true, cache=nothing)
+
+SIMD-accelerated adaptive 1D Tanh-Sinh integration using `LoopVectorization.@turbo`.
+This is an opt-in fast path: `f` must be compatible with `@turbo`, and floating-point
+rounding may differ slightly from `adaptive_integrate_1D` due to reassociated reductions.
+"""
+function adaptive_integrate_1D_avx(::Type{T}, f::F, a_T::T, b_T::T,
+    rtol_T::T, atol_T::T, max_levels::Int,
+    warn::Bool, cache1d::_Adaptive1DCache{T}) where {T<:Real,F}
+    Δx, x₀ = _midpoint_radius(a_T, b_T)
+    half = _half(T)
+    h = cache1d.tm * half
+    w0 = _half_pi(T)
+    s_total = w0 * f(x₀)
+
+    @inbounds for i in 1:2
+        xk = cache1d.initial_x[i]
+        s_total += cache1d.initial_w[i] * (f(Δx * xk + x₀) + f(x₀ - Δx * xk))
+    end
+
+    old_res = Δx * h * s_total
+    err_est = zero(T)
+    for level in 1:max_levels
+        h *= half
+        s_new = zero(T)
+        x_level = cache1d.xs[level]
+        w_level = cache1d.ws[level]
+        @turbo for i in eachindex(x_level)
+            xk = x_level[i]
+            s_new += w_level[i] * (f(Δx * xk + x₀) + f(x₀ - Δx * xk))
+        end
+
+        s_total += s_new
+        new_res = Δx * h * s_total
+        err_est = abs(new_res - old_res)
+
+        if err_est <= _error_target(new_res, rtol_T, atol_T)
+            return new_res
+        end
+        old_res = new_res
+    end
+    if warn && max_levels > 0
+        @warn "adaptive_integrate_1D_avx reached max_levels without meeting the requested tolerance." max_levels estimated_error = err_est target = _error_target(old_res, rtol_T, atol_T) value = old_res rtol = rtol_T atol = atol_T
+    end
+    return old_res
+end
+
+function adaptive_integrate_1D_avx(::Type{T}, f::F, a, b;
+    rtol=nothing, atol::Real=0, max_levels::Int=16,
+    warn::Bool=true, cache=nothing) where {T<:Real,F}
+    a_T, b_T = T(a), T(b)
[…]
