### 2026-04-06 fcb6bd1
Add experimental integration benchmarks
