+++ test/experimental_adaptive_3d.jl
@@ -1,0 +1,493 @@
+using BenchmarkTools
+using FastTanhSinhQuadrature
+using LoopVectorization
+using StaticArrays
+using StrideArrays
+
+struct PackageStyleLevelData{X,W,XP,XM,YP,YM,ZP,ZM,O}
+    x::X
+    w::W
+    xp::XP
+    xm::XM
+    yp::YP
+    ym::YM
+    zp::ZP
+    zm::ZM
+    owner::O
+end
+
+struct PackageStyleVariant{T,L}
+    name::String
+    tm::T
+    initial_x::NTuple{2,T}
+    initial_w::NTuple{2,T}
+    levels::L
+end
+
+@inline to_svector(::Val{N}, v::AbstractVector{T}) where {N,T} =
+    SVector{N,T}(ntuple(i -> v[i], Val(N)))
+
+@inline level_node_count(::Val{L}) where {L} = 1 << (L + 1)
+
+function build_level_data(::Val{:vector}, ::Val{L}, x_src, w_src) where {L}
+    T = eltype(x_src)
+    n = length(x_src)
+    x = collect(x_src)
+    w = collect(w_src)
+    xp = Vector{T}(undef, n)
+    xm = Vector{T}(undef, n)
+    yp = Vector{T}(undef, n)
+    ym = Vector{T}(undef, n)
+    zp = Vector{T}(undef, n)
+    zm = Vector{T}(undef, n)
+    return PackageStyleLevelData(x, w, xp, xm, yp, ym, zp, zm, nothing)
+end
+
+function build_level_data(::Val{:ptr_vector}, ::Val{L}, x_src, w_src) where {L}
+    T = eltype(x_src)
+    n = length(x_src)
+    x = collect(x_src)
+    w = collect(w_src)
+    xp = Vector{T}(undef, n)
+    xm = Vector{T}(undef, n)
+    yp = Vector{T}(undef, n)
+    ym = Vector{T}(undef, n)
+    zp = Vector{T}(undef, n)
+    zm = Vector{T}(undef, n)
+    owner = (x, w, xp, xm, yp, ym, zp, zm)
+    return PackageStyleLevelData(
+        PtrArray(x),
+        PtrArray(w),
+        PtrArray(xp),
+        PtrArray(xm),
+        PtrArray(yp),
+        PtrArray(ym),
+        PtrArray(zp),
+        PtrArray(zm),
+        owner,
+    )
+end
+
+function build_level_data(::Val{:stride}, ::Val{L}, x_src, w_src) where {L}
+    T = eltype(x_src)
+    n = length(x_src)
+    x = StrideArray(collect(x_src))
+    w = StrideArray(collect(w_src))
+    xp = StrideArray{T}(undef, (n,))
+    xm = StrideArray{T}(undef, (n,))
+    yp = StrideArray{T}(undef, (n,))
+    ym = StrideArray{T}(undef, (n,))
+    zp = StrideArray{T}(undef, (n,))
+    zm = StrideArray{T}(undef, (n,))
+    return PackageStyleLevelData(x, w, xp, xm, yp, ym, zp, zm, nothing)
+end
+
+function build_level_data(::Val{:ptr_stride}, ::Val{L}, x_src, w_src) where {L}
+    T = eltype(x_src)
+    n = length(x_src)
+    x = StrideArray(collect(x_src))
+    w = StrideArray(collect(w_src))
+    xp = StrideArray{T}(undef, (n,))
+    xm = StrideArray{T}(undef, (n,))
+    yp = StrideArray{T}(undef, (n,))
+    ym = StrideArray{T}(undef, (n,))
+    zp = StrideArray{T}(undef, (n,))
[…]
