+++ test/experimental_performance3D_stride.jl
@@ -1,0 +1,295 @@
+using FastTanhSinhQuadrature
+using StaticArrays
+using BenchmarkTools
+using LoopVectorization
+using StrideArrays
+
+struct Integration3DVectorCache{T}
+    xp::Vector{T}
+    xm::Vector{T}
+    yp::Vector{T}
+    ym::Vector{T}
+    zp::Vector{T}
+    zm::Vector{T}
+end
+
+function Integration3DVectorCache(::Type{T}, n::Int) where {T}
+    return Integration3DVectorCache{T}(
+        Vector{T}(undef, n),
+        Vector{T}(undef, n),
+        Vector{T}(undef, n),
+        Vector{T}(undef, n),
+        Vector{T}(undef, n),
+        Vector{T}(undef, n),
+    )
+end
+
+struct Integration3DStrideCache{A}
+    xp::A
+    xm::A
+    yp::A
+    ym::A
+    zp::A
+    zm::A
+end
+
+function Integration3DStrideCache(::Type{T}, n::Int) where {T}
+    A = typeof(StrideArray{T}(undef, (n,)))
+    return Integration3DStrideCache{A}(
+        StrideArray{T}(undef, (n,)),
+        StrideArray{T}(undef, (n,)),
+        StrideArray{T}(undef, (n,)),
+        StrideArray{T}(undef, (n,)),
+        StrideArray{T}(undef, (n,)),
+        StrideArray{T}(undef, (n,)),
+    )
+end
+
+struct Integration3DPtrCache{A}
+    xp::A
+    xm::A
+    yp::A
+    ym::A
+    zp::A
+    zm::A
+end
+
+Integration3DPtrCache(cache) =
+    Integration3DPtrCache(
+        PtrArray(cache.xp),
+        PtrArray(cache.xm),
+        PtrArray(cache.yp),
+        PtrArray(cache.ym),
+        PtrArray(cache.zp),
+        PtrArray(cache.zm),
+    )
+
+struct KernelVariant{X,W,C}
+    name::String
+    x::X
+    w::W
+    cache::C
+end
+
+@inline to_svector(::Val{N}, v::AbstractVector{T}) where {N,T} =
+    SVector{N,T}(ntuple(i -> v[i], Val(N)))
+
+function _integrate3D_core(
+    f::F, low::SVector{3,T}, up::SVector{3,T}, h::T,
+    x::AbstractVector{T}, w::AbstractVector{T},
+    xp::AbstractVector{T}, xm::AbstractVector{T},
+    yp::AbstractVector{T}, ym::AbstractVector{T},
+    zp::AbstractVector{T}, zm::AbstractVector{T},
+) where {T<:Real,F}
+    Δx, x₀ = (up[1] - low[1]) / 2, (up[1] + low[1]) / 2
+    Δy, y₀ = (up[2] - low[2]) / 2, (up[2] + low[2]) / 2
+    Δz, z₀ = (up[3] - low[3]) / 2, (up[3] + low[3]) / 2
+    n = length(x)
+
+    @turbo for i in 1:n
+        xi = x[i]
+        vx = Δx * xi
+        vy = Δy * xi
+        vz = Δz * xi
+        xp[i] = x₀ + vx
+        xm[i] = x₀ - vx
+        yp[i] = y₀ + vy
+        ym[i] = y₀ - vy
+        zp[i] = z₀ + vz
+        zm[i] = z₀ - vz
+    end
+
+    w₀ = T(π) / 2
+    w₀² = w₀ * w₀
+    w₀³ = w₀² * w₀
+
+    total_sum = w₀³ * f(x₀, y₀, z₀)
+
[…]

[…]
