### 2026-04-06 5782799
Simplify tensor coordinate expressions
