+++ src/integrate3D.jl
@@ -79,13 +79,10 @@ end
         @turbo for i in eachindex(x)
             wi = w[i]
             xi = x[i]
-            dx = Δx * xi
-            dy = Δy * xi
-            dz = Δz * xi
             total_sum += wi * w₀² * (
-                f(x₀ + dx, y₀, z₀) + f(x₀ - dx, y₀, z₀) +
-                f(x₀, y₀ + dy, z₀) + f(x₀, y₀ - dy, z₀) +
-                f(x₀, y₀, z₀ + dz) + f(x₀, y₀, z₀ - dz)
+                f(Δx * xi + x₀, y₀, z₀) + f(x₀ - Δx * xi, y₀, z₀) +
+                f(x₀, Δy * xi + y₀, z₀) + f(x₀, y₀ - Δy * xi, z₀) +
+                f(x₀, y₀, Δz * xi + z₀) + f(x₀, y₀, z₀ - Δz * xi)
             )
         end
 
@@ -95,14 +92,13 @@ end
             wj = w[j]
             xi = x[i]
             xj = x[j]
-            dx = Δx * xi
-            dy_i = Δy * xi
-            dy = Δy * xj
-            dz = Δz * xj
             plane_sum += wi * wj * w₀ * (
-                f(x₀ + dx, y₀ + dy, z₀) + f(x₀ - dx, y₀ + dy, z₀) + f(x₀ + dx, y₀ - dy, z₀) + f(x₀ - dx, y₀ - dy, z₀) +
-                f(x₀ + dx, y₀, z₀ + dz) + f(x₀ - dx, y₀, z₀ + dz) + f(x₀ + dx, y₀, z₀ - dz) + f(x₀ - dx, y₀, z₀ - dz) +
-                f(x₀, y₀ + dy_i, z₀ + dz) + f(x₀, y₀ - dy_i, z₀ + dz) + f(x₀, y₀ + dy_i, z₀ - dz) + f(x₀, y₀ - dy_i, z₀ - dz)
+                f(Δx * xi + x₀, Δy * xj + y₀, z₀) + f(x₀ - Δx * xi, Δy * xj + y₀, z₀) +
+                f(Δx * xi + x₀, y₀ - Δy * xj, z₀) + f(x₀ - Δx * xi, y₀ - Δy * xj, z₀) +
+                f(Δx * xi + x₀, y₀, Δz * xj + z₀) + f(x₀ - Δx * xi, y₀, Δz * xj + z₀) +
+                f(Δx * xi + x₀, y₀, z₀ - Δz * xj) + f(x₀ - Δx * xi, y₀, z₀ - Δz * xj) +
+                f(x₀, Δy * xi + y₀, Δz * xj + z₀) + f(x₀, y₀ - Δy * xi, Δz * xj + z₀) +
+                f(x₀, Δy * xi + y₀, z₀ - Δz * xj) + f(x₀, y₀ - Δy * xi, z₀ - Δz * xj)
             )
         end
 
@@ -114,14 +110,11 @@ end
             xi = x[i]
             xj = x[j]
             xk = x[k]
-            dx = Δx * xi
-            dy = Δy * xj
-            dz = Δz * xk
             octant_sum += wi * wj * wk * (
-                f(x₀ + dx, y₀ + dy, z₀ + dz) + f(x₀ - dx, y₀ + dy, z₀ + dz) +
-                f(x₀ + dx, y₀ - dy, z₀ + dz) + f(x₀ - dx, y₀ - dy, z₀ + dz) +
-                f(x₀ + dx, y₀ + dy, z₀ - dz) + f(x₀ - dx, y₀ + dy, z₀ - dz) +
-                f(x₀ + dx, y₀ - dy, z₀ - dz) + f(x₀ - dx, y₀ - dy, z₀ - dz)
[…]
