### 2026-04-06 5872487
v2.1.0: optimized 3D adaptive AVX paths
