### 2026-04-06 #10076 @svretina
DOI: https://zenodo.org/records/19420466
Version: v2.1.0

DOI: https://zenodo.org/records/19420465 (for all versions)

affiliations get simplified in zenodo but are the same institutes as mentioned in the paper.
