### 2026-04-16 #10076 @danielskatz
👋 @svretina - as track chair, I've proofread this submission, and see one small change that is needed, in https://github.com/svretina/FastTanhSinhQuadrature.jl/pull/37 - please merge this, then we can move forward.
