### 2026-04-16 #10076 @danielskatz
Congratulations to @svretina (Stamatis Vretinaris) on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you.
